Add location of each pair of duplication in bed
Some tiny codes were added in 'purge_dup.c' for taking locus information which are pairing with each duplication (HAPLOTIGS or OVLP) in bed file.
I'm not good in C, I do not know it is appropriately revised for this objectives yet.
This revision have to be confirmed that whether it is available.